+++ semver_internal_test.go
@@ -10,26 +10,9 @@ func BenchmarkIsValidByParse(b *testing.B) {
 	}
 }
 
-func BenchmarkIsValidPrefixByParse(b *testing.B) {
-	test := "semver0.1.0-alpha.24+sha.19031c2.darwin.amd64"
-
-	for range b.N {
-		_ = isValidPrefixByParse(test, "semver")
-	}
-}
-
 // isValidByParse is the old implementation of the validation function.
 func isValidByParse(s string) bool {
 	if _, err := parse(s); err != nil {
-		return false
-	}
-
-	return true
-}
-
-// isValidPrefixByParse is the old implementation of the validation function.
-func isValidPrefixByParse(s string, prefixes ...string) bool {
-	if _, err := parse(s, prefixes...); err != nil {
 		return false
 	}